requirements.txt
PyYAML>=4.2b1
certifi==2018.4.16
chardet==3.0.4
idna==2.6
pytz==2018.4
requests>=2.20.0
urllib3>=1.24.2